--- Springboot/src/main/java/com/centralstudenthub/error/AllCoursesAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.centralstudenthub.error;
-
-public class AllCoursesAlreadyExistsException extends Exception{
-    public AllCoursesAlreadyExistsException(String message) {
-        super(message);
-    }
-}
