--- Springboot/src/main/java/com/centralstudenthub/error/ConflictException.java
@@ -1,7 +1,0 @@
-package com.centralstudenthub.error;
-
-public class ConflictException extends Error {
-    public ConflictException(String message) {
-        super(message);
-    }
-}
