--- Springboot/src/main/java/com/centralstudenthub/error/CourseAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.centralstudenthub.error;
-
-public class CourseAlreadyExistsException extends Exception{
-    public CourseAlreadyExistsException(String message) {
-        super(message);
-    }
-}
