--- Springboot/src/main/java/com/centralstudenthub/error/CourseNotFoundException.java
@@ -1,7 +1,0 @@
-package com.centralstudenthub.error;
-
-public class CourseNotFoundException extends Exception{
-    public CourseNotFoundException(String message) {
-        super(message);
-    }
-}
