+++ Springboot/src/main/java/com/centralstudenthub/exception/AllCoursesAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.centralstudenthub.exception;
+
+public class AllCoursesAlreadyExistsException extends Exception {
+    public AllCoursesAlreadyExistsException(String message) {
+        super(message);
+    }
+}
