+++ Springboot/src/main/java/com/centralstudenthub/exception/ConflictException.java
@@ -1,0 +1,7 @@
+package com.centralstudenthub.exception;
+
+public class ConflictException extends Exception {
+    public ConflictException(String message) {
+        super(message);
+    }
+}
