+++ Springboot/src/main/java/com/centralstudenthub/exception/CourseAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.centralstudenthub.exception;
+
+public class CourseAlreadyExistsException extends Exception {
+    public CourseAlreadyExistsException(String message) {
+        super(message);
+    }
+}
