+++ Springboot/src/main/java/com/centralstudenthub/exception/CourseNotFoundException.java
@@ -1,0 +1,7 @@
+package com.centralstudenthub.exception;
+
+public class CourseNotFoundException extends Exception {
+    public CourseNotFoundException(String message) {
+        super(message);
+    }
+}
