+++ Springboot/src/main/java/com/centralstudenthub/service/CoursePrerequisiteService.java
@@ -3,9 +3,9 @@ package com.centralstudenthub.service;
 import com.centralstudenthub.entity.student_profile.course.Course;
 import com.centralstudenthub.entity.student_profile.course.course_prerequisites.CoursePrerequisite;
 import com.centralstudenthub.entity.student_profile.course.course_prerequisites.CoursePrerequisiteId;
-import com.centralstudenthub.error.ConflictException;
-import com.centralstudenthub.error.CourseAlreadyExistsException;
-import com.centralstudenthub.error.CourseNotFoundException;
+import com.centralstudenthub.exception.ConflictException;
+import com.centralstudenthub.exception.CourseAlreadyExistsException;
+import com.centralstudenthub.exception.CourseNotFoundException;
 import com.centralstudenthub.repository.CoursePrerequisiteRepository;
 import com.centralstudenthub.repository.CourseRepository;
 import jakarta.transaction.Transactional;
