+++ Springboot/src/main/java/com/centralstudenthub/service/CourseService.java
@@ -3,9 +3,9 @@ package com.centralstudenthub.service;
 import com.centralstudenthub.Model.Request.CourseRequest;
 import com.centralstudenthub.Model.Response.CourseResponse;
 import com.centralstudenthub.entity.student_profile.course.Course;
-import com.centralstudenthub.error.AllCoursesAlreadyExistsException;
-import com.centralstudenthub.error.CourseAlreadyExistsException;
-import com.centralstudenthub.error.CourseNotFoundException;
+import com.centralstudenthub.exception.AllCoursesAlreadyExistsException;
+import com.centralstudenthub.exception.CourseAlreadyExistsException;
+import com.centralstudenthub.exception.CourseNotFoundException;
 import com.centralstudenthub.repository.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
